fix: resolve plugin deps by registry name, not dir name
This is a rather large way to go about doing this - in essence this is a re-write of the whole plugin loader execution flow, which does a pass of resolving configs *first*, and then once registry names are known, goes back and does a loading sweep.

Another change is that dependency order is resolved *only* at plugin *start* time, rather than at load time. Plugins that depend on others should resolve their dependencies during their `init` block. Doing this cuts off one extra way for plugins to "leak" when they should be unloaded.

Finally, callback hell has been replaced with `Promise` hell. Unfortunately, Haxe does not natively support `async`/`await`, so we move from nesting callbacks, to chaining `then`, which is... marginally better?

The main good part about this is being able to use `Promise.all()` and related constructs, along with preparing the plugin manager to support a potential soft-dependency system that'd use promises within `init` to gain access to another plugin's data (ideally an intentionally exposed API.)

Finally, there's an unresolved bug currently where `builder` fails to initialise. I believe this is due to it relying on image assets for icons which Electron appears to require to be loaded immediately, or will throw an error.